+++ client/src/streams/pubsub/subscriber.rs
@@ -1,6 +1,7 @@
 use super::states::{SubscriberWantsDecoder, SubscriberWantsOpen};
 use crate::connection::{ClientConnection, SharedConnection};
-use crate::keep_alive::{AttemptFut, KeepAlive};
+use crate::keep_alive::pubsub::KeepAlive;
+use crate::keep_alive::AttemptFut;
 use crate::streams::aliases::Decomp;
 use crate::streams::handle_reply;
 use crate::traits::{KeepAliveStream, Open, Operations, Retain, TryIntoU64};
@@ -74,7 +75,7 @@ where
     D: MessageDecoder<Item> + Send + Unpin,
     Item: Send + Unpin,
 {
-    type Output = KeepAlive<Subscriber<D, Item>, Item>;
+    type Output = KeepAlive<Subscriber<D, Item>>;
 
     async fn open(self) -> Result<Self::Output> {
         let topic = TopicName::try_from(self.state.common.topic.as_str())?;
@@ -125,7 +126,7 @@ where
         headers: SubscriberPayload,
         decoder: D,
         decompression: Option<Decomp>,
-    ) -> Result<KeepAlive<Self, Item>> {
+    ) -> Result<KeepAlive<Self>> {
         let lock = client.connection.lock().await;
         let stream = Self::open_stream(lock, headers.clone()).await?;